[Add] 공통적으로 사용되는 js 파일 구조화 #82
공통적으로 헤더와 푸터를 부르는 기능을 common.js로 따로 분류 하였습니다.
