+++ client/index.html
@@ -7,7 +7,6 @@
 		<link rel="shortcut icon" href="https://res.kurly.com/favicon.ico" type="res/ico" />
 		<link rel="stylesheet" href="/css/main.css" />
 		<script src="/js/main.js" type="module"></script>
-
 		<title>칼리 - 마켓칼리/뷰티칼리</title>
 	</head>
 	<body>
