+++ client/js/common/index.js
@@ -1,0 +1,1 @@
+export * from './common.js';
